Moved contents of remote_data_folder.rs to arguments.rs
