+++ crates/modelardb_common/src/lib.rs
@@ -20,4 +20,3 @@ pub mod macros;
 pub mod schemas;
 pub mod types;
 pub mod arguments;
-pub mod remote_data_folder;
