+++ crates/modelardb_manager/src/main.rs
@@ -21,8 +21,8 @@ use std::sync::Arc;
 
 use modelardb_common::arguments::{
     argument_to_remote_object_store, collect_command_line_arguments,
+    validate_remote_data_folder_from_argument,
 };
-use modelardb_common::remote_data_folder::validate_remote_data_folder_from_argument;
 use sqlx::postgres::PgConnectOptions;
 use sqlx::{ConnectOptions, PgConnection};
 use tokio::runtime::Runtime;
