+++ crates/modelardb_server/src/main.rs
@@ -33,8 +33,8 @@ use datafusion::execution::context::{SessionConfig, SessionContext, SessionState
 use datafusion::execution::runtime_env::RuntimeEnv;
 use modelardb_common::arguments::{
     argument_to_remote_object_store, collect_command_line_arguments,
+    validate_remote_data_folder_from_argument,
 };
-use modelardb_common::remote_data_folder::validate_remote_data_folder_from_argument;
 use object_store::{local::LocalFileSystem, ObjectStore};
 use once_cell::sync::Lazy;
 use tokio::runtime::Runtime;
